Add files via upload
This is a replacement website for the current one i have hosted. This needs a lot more work until it could be called completed. I am doing it all from scratch and wont be using a template of any kind. This website may look far different when it is completed. Also please keep in mind if you see any colors that look ugly or formatting that looks ugly, This is in development and until it is a polished product ready for action it may not look good because i cant always tell where my divs are so i find filling them with color makes it easy to see how they are being manipulated. Every line of code is by me and me alone, I expect much more progress to be made after i get done with college this semester.